requirements.txt
python-dotenv
yfinance
pymongo